requirements.txt
python-telegram-bot==20.7
requests==2.31.0
flask==3.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0